+++ site/src/Report/Cashflow/CashflowReportParams.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Report\Cashflow;
+
+use App\Entity\Company;
+
+final class CashflowReportParams
+{
+    public function __construct(
+        public readonly Company $company,
+        public readonly string $group,
+        public readonly \DateTimeImmutable $from,
+        public readonly \DateTimeImmutable $to,
+    ) {
+    }
+}
