+++ site/templates/finance/report/cashflow.html.twig
@@ -1,0 +1,1 @@
+{% extends 'report/cashflow.html.twig' %}
